\name{sim_data}
\alias{sim_data}
\title{Simulated data}
\description{
Simulated data for Bayesian robust multivariate linear regression with incomplete response.
}
\usage{
sim_data
}
\format{
Simulated data with 50 observations, and each observation has 2 response components and 1 predictor:
\code{yr1}, \code{yr2}, \code{x}.
}



\examples{
# sim_data is generated by the following code
set.seed(1)

library(mvtnorm)
n <- 50
X <- cbind(1, as.matrix(rnorm(n, 0, 1)))
beta <- matrix(c(0.8, 0.4, 0.5, 0.3), nrow=2, ncol=2)
var <- matrix(c(1, 0.6, 0.6, 1), 2, 2)
error <- matrix(0, nrow=n, ncol=2)
w <- rgamma(n, shape=2, rate=2)
for (i in 1:n) {
  error[i, ] <- rmvnorm(1, rep(0, nrow(var)), var / w[i])
}
yobs <- X \%*\% beta + error
sim_data <- cbind(yobs, X[, 2])
colnames(sim_data) <- c("yr1", "yr2", "x")
}
